Update location attrs to double and validate min and max values

+++ src/main/kotlin/com/eduardo/rdguez/zssn/domain/Location.kt
@@ -15,8 +15,8 @@ class Location(
   var id: Long? = null,
 
   @Column(name = "latitude")
-  var latitude: Int,
+  var latitude: Double,
 
   @Column(name = "longitude")
-  var longitude: Int,
+  var longitude: Double,
 )
